Rename PasswordHashCreator -> PasswordHasher and add test for multithreading
